app/requirements.txt
Flask
mysql-connector-python
joblib
scikit-learn==1.3.2

